Add test for feature record
