+++ tests/test_data_set_features.py
@@ -1,5 +1,4 @@
 import pytest
-import ipfx.feature_record as fr
 import ipfx.data_set_features as dsft
 from ipfx.error import FeatureError
 
@@ -44,17 +43,4 @@ def test_select_subthreshold_min_amplitude():
     assert min_amp == -100
 
 
-def test_nan_get():
-
-    a = {}
-    v = fr.nan_get(a, 'fish')
-    assert v == None
-
-    a = { 'fish': 1 }
-    v = fr.nan_get(a, 'fish')
-    assert v == 1
-
-    a = { 'fish': float("nan") }
-    v = fr.nan_get(a, 'fish')
-    assert v == None
